+++ src/main/java/com/coloronme/admin/domain/consult/dto/response/ConsultUserResponseDto.java
@@ -14,7 +14,7 @@ public class ConsultUserResponseDto {
     private String nickname;
     private String email;
     private LocalDateTime personalDate;
-    private Long personalColorId;
+    private int personalColorId;
     private String age;
     private Gender gender;
     private String consultContent;
