+++ src/main/java/com/coloronme/admin/domain/consultant/repository/ConsultantRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ConsultantRepository extends JpaRepository<Consultant, Long> {
+public interface ConsultantRepository extends JpaRepository<Consultant, Integer> {
     Optional<Consultant> findByEmail(String email);
 }
